Keep the styles we did not write together
They are somebody else's cartography, carried along as the JSON it was
published as, which is the one thing they have in common and the one thing
which sets them apart from the basemap. Naming the module after either
source would have collided with the schema constants, and this way they
leave in one piece when they leave.
