+++ walkers/src/style/mod.rs
@@ -1,4 +1,5 @@
 pub mod basemap;
+mod third_party;
 
 use color::Rgba8;
 use egui::Color32;
@@ -20,37 +21,6 @@ use crate::expression::Context;
 #[derive(Deserialize, Default)]
 pub struct Style {
     pub layers: Vec<Layer>,
-}
-
-impl Style {
-    /// Style based on Protomaps Dark flavour. Requires Protomaps source.
-    ///
-    /// <https://docs.protomaps.com/basemaps/flavors>
-    pub fn protomaps_dark() -> Self {
-        let style_json = include_str!("../../assets/protomaps-dark.json");
-        serde_json::from_str(style_json).expect("failed to parse style JSON")
-    }
-
-    /// Style based on Protomaps Dark Vis flavour. Requires Protomaps source.
-    ///
-    /// <https://docs.protomaps.com/basemaps/flavors>
-    pub fn protomaps_dark_vis() -> Self {
-        let style_json = include_str!("../../assets/protomaps-dark-vis.json");
-        serde_json::from_str(style_json).expect("failed to parse style JSON")
-    }
-
-    /// Style based on Protomaps Light flavour. Requires Protomaps source.
-    ///
-    /// <https://docs.protomaps.com/basemaps/flavors>
-    pub fn protomaps_light() -> Self {
-        let style_json = include_str!("../../assets/protomaps-light.json");
-        serde_json::from_str(style_json).expect("failed to parse style JSON")
-    }
-
-    pub fn openfreemap_bright() -> Self {
-        let style_json = include_str!("../../assets/openfreemap-bright.json");
-        serde_json::from_str(style_json).expect("failed to parse style JSON")
-    }
 }
 
 /// Which layer, or layers, of a vector tile a style layer draws from.
